+++ src/middleware/index.ts
@@ -1,16 +1,1 @@
-import { UiState, InstantSearch } from '../types';
-
-export type MiddlewareDefinition = {
-  onStateChange(options: { uiState: UiState }): void;
-  subscribe(): void;
-  unsubscribe(): void;
-};
-
-export type MiddlewareOptions = {
-  instantSearchInstance: InstantSearch;
-};
-
-export type Middleware = (options: MiddlewareOptions) => MiddlewareDefinition;
-
-export { createRouter, RouterProps } from './createRouter';
 export { createInsightsMiddleware } from './insights';
